requirements.txt
aiohttp==3.7.4.post0
amqp==5.0.6
asgiref==3.3.1
async-timeout==3.0.1
attrs==21.2.0
billiard==3.6.4.0
boto3==1.17.94
botocore==1.20.94
celery==5.1.2
certifi==2021.5.30
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
dj-database-url==0.5.0
Django==3.1.5
django-common-helpers==0.9.2
django-cors-headers==3.7.0
django-crispy-forms==1.12.0
django-cron==0.5.1
django-filter==2.4.0
django-heroku==0.3.1
django-resized==0.3.11
django-storages==1.11.1
django-user-agents==0.4.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.7.2
geoip2==4.2.0
gunicorn==20.1.0
idna==2.10
jmespath==0.10.0
kombu==5.1.0
Markdown==3.3.3
maxminddb==2.0.3
multidict==5.1.0
Pillow==8.1.0
prompt-toolkit==3.0.20
psycopg2==2.8.6
PyJWT==2.0.1
python-dateutil==2.8.1
python-decouple==3.4
pytz==2020.5
PyYAML==5.4.1
redis==3.5.3
requests==2.25.1
s3transfer==0.4.2
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
ua-parser==0.10.0
urllib3==1.26.6
user-agents==2.2.0
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.2.0
yarl==1.6.3
